[16.0][FIX] document_page: Don't change last history name and summary when creating new revision of page.
Fixes an issue where editing a document.page and changing name and/or summary
(to create a new revision) also changes name and summary of the history_head.

draft_name and draft_summary are now computed fields (based on history_head)
but have set the inverse without actually setting the value, to allow setting
them when creating a new draft page.

The issue is prominently evident when using document_page_approval where
the name/summary for the new draft also change the last approved version
(aka history_head) which should not happen.